requirements.txt
# Generic Libraries
openpyxl==3.1.5
python-dotenv==1.0.0
pandas==1.3.3
numpy==1.24.4
#Spire.Doc==13.1.0
markdown==3.7.0

# Web Scraping and Keyword Extraction
#pytrends==4.9.2
googlesearch-python==1.3.0
beautifulsoup4==4.12.3
nltk==3.8.0
requests==2.32.3

# AI Content Generation
transformers==4.30.2
openai==0.27.0
torch==2.0.0

# Text Analysis
textstat==0.7.3
spacy==3.5.0
textblob==0.19.0
#sentence_transformers==3.4.0

# Additional if needed:
#python-wordpress-xmlrpc #(if you're using WordPress)